refactor: move util packages
