+++ handlers/admin/adminIPBanPage.go
@@ -6,7 +6,7 @@ import (
 	corecommon "github.com/arran4/goa4web/core/common"
 	common "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/netutil"
+	"github.com/arran4/goa4web/internal/utils/netutil"
 	"log"
 	"net/http"
 	"strings"
