+++ handlers/auth/forgotPassword.go
@@ -11,7 +11,7 @@ import (
 	"github.com/arran4/goa4web/core/templates"
 	common "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/emailutil"
+	"github.com/arran4/goa4web/internal/utils/emailutil"
 )
 
 func ForgotPasswordPage(w http.ResponseWriter, r *http.Request) {
