+++ handlers/blogs/blogsBlogReplyPage.go
@@ -7,8 +7,8 @@ import (
 	"github.com/arran4/goa4web/core"
 	hcommon "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
-	"github.com/arran4/goa4web/internal/emailutil"
-	searchutil "github.com/arran4/goa4web/internal/searchutil"
+	"github.com/arran4/goa4web/internal/utils/emailutil"
+	searchutil "github.com/arran4/goa4web/internal/utils/searchutil"
 	"github.com/gorilla/mux"
 	"log"
 	"net/http"
