+++ handlers/news/newsPostPage.go
@@ -19,9 +19,9 @@ import (
 	hcommon "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
 	email "github.com/arran4/goa4web/internal/email"
-	"github.com/arran4/goa4web/internal/emailutil"
 	notif "github.com/arran4/goa4web/internal/notifications"
-	searchutil "github.com/arran4/goa4web/internal/searchutil"
+	"github.com/arran4/goa4web/internal/utils/emailutil"
+	searchutil "github.com/arran4/goa4web/internal/utils/searchutil"
 	"github.com/arran4/goa4web/runtimeconfig"
 )
