+++ handlers/news/searchResultNewsActionPage.go
@@ -11,7 +11,7 @@ import (
 	"github.com/arran4/goa4web/core/templates"
 	hcommon "github.com/arran4/goa4web/handlers/common"
 	db "github.com/arran4/goa4web/internal/db"
-	searchutil "github.com/arran4/goa4web/internal/searchutil"
+	searchutil "github.com/arran4/goa4web/internal/utils/searchutil"
 )
 
 func SearchResultNewsActionPage(w http.ResponseWriter, r *http.Request) {
