+++ handlers/search/search_test.go
@@ -5,7 +5,7 @@ import (
 	"database/sql"
 	"errors"
 	dbpkg "github.com/arran4/goa4web/internal/db"
-	searchutil "github.com/arran4/goa4web/internal/searchutil"
+	searchutil "github.com/arran4/goa4web/internal/utils/searchutil"
 	"net/http"
 	"net/http/httptest"
 	"testing"
