+++ handlers/user/userEmailPage.go
@@ -19,7 +19,7 @@ import (
 	db "github.com/arran4/goa4web/internal/db"
 
 	"github.com/arran4/goa4web/internal/email"
-	"github.com/arran4/goa4web/internal/emailutil"
+	"github.com/arran4/goa4web/internal/utils/emailutil"
 
 	"github.com/arran4/goa4web/runtimeconfig"
 )
