+++ internal/utils/searchutil/insert.go
